+++ WebApplication/src/store/modules/labels.js
@@ -1,9 +1,12 @@
 import api from "../../orthancApi"
+import store from "../../store"
+
 
 ///////////////////////////// STATE
 const state = () => ({
     allLabels: [],    // all labels in Orthanc
-    loaded: false
+    loaded: false,
+    canShowStudiesWithoutLabels: false
 })
 
 ///////////////////////////// GETTERS
@@ -18,6 +21,9 @@ const mutations = {
     },
     setLoaded(state) {
         state.loaded = true;
+    },
+    setCanShowStudiesWithoutLabels(state) {
+        state.canShowStudiesWithoutLabels = true;
     }
 }
 
@@ -27,6 +33,20 @@ const actions = {
     async refresh({ commit }) {
         const allLabels = await api.loadAllLabels();
         commit('setLabels', { allLabels: allLabels });
+
+        let showStudiesWithoutLabels = false;
+        if (store.state.configuration.system.ApiVersion >= 30 && allLabels.length > 0)
+        {
+            if (store.state.configuration.userProfile) { // only the users who have access to all labels can see this option !
+                showStudiesWithoutLabels = store.state.configuration.userProfile["authorized-labels"].length == 1 && store.state.configuration.userProfile["authorized-labels"][0] == "*";
+            } else {
+                showStudiesWithoutLabels = true;
+            }
+            if (showStudiesWithoutLabels) {
+                commit('setCanShowStudiesWithoutLabels');
+            }
+        }
+
         commit('setLoaded');
     }
 }